Move ConfigurationAcceptor as top level class
