+++ dd-java-agent/agent-debugger/src/main/java/com/datadog/debugger/agent/ConfigurationAcceptor.java
@@ -1,0 +1,13 @@
+package com.datadog.debugger.agent;
+
+import com.datadog.debugger.probe.ProbeDefinition;
+import java.util.Collection;
+
+public interface ConfigurationAcceptor {
+  enum Source {
+    REMOTE_CONFIG,
+    EXCEPTION
+  }
+
+  void accept(Source source, Collection<? extends ProbeDefinition> definitions);
+}
